chore: separate library and binary crate
